Refactor basket
Create a Basket model in core and move basket logic
from front to core so that a basket can be used without
any need for shuup.front. Unlike the front basket, the core
basket does not use sessions for basket retrieval but is
instead reliant on the caller to pass in the `basket.key`.
