+++ shuup/core/basket/order_creator.py
@@ -5,6 +5,8 @@
 #
 # This source code is licensed under the OSL-3.0 license found in the
 # LICENSE file in the root directory of this source tree.
-from shuup.core.basket.command_dispatcher import BasketCommandDispatcher
+from shuup.core.order_creator import OrderCreator
 
-BasketCommandDispatcher = BasketCommandDispatcher  # noqa
+
+class BasketOrderCreator(OrderCreator):
+    pass
